add H2

+++ src/main/java/com/example/java/Models/Order.java
@@ -3,6 +3,6 @@ package com.example.java.Models;
 import java.text.DecimalFormat;
 import java.util.Date;
 
-public record Order(Integer orderId, Integer clientId, Float price, Date orderTime) {
+public record Order(Integer orderId, Integer clientId, Integer price, Date orderTime) {
 
 }
